+++ qomputing-frontend/utils/constants/footer.ts
@@ -1,0 +1,41 @@
+export const footerContent = {
+  heading: 'Qomputing',
+  subheading: 'Empowering the future of computation with our comprehensive quantum development environment. Join us on the quantum journey.',
+  newsletter: 'Join Our Newsletter',
+  subletter: 'Stay updated with the latest in quantum computing.',
+  subscribe: 'Subscribe',
+  copyright: '© {new Date().getFullYear()} Qomputing. All rights reserved.',
+  terms: 'Terms of Service',
+  privacy: 'Privacy Policy',
+  cookies: 'Cookies'
+};
+
+export const footerLinks = [
+  {
+    title: "Products",
+    links: [
+      { label: "Qomputing IDE", href: "#" },
+      { label: "Qomputing Simulator", href: "#" },
+      { label: "Quantum Tools", href: "#" },
+      { label: "Pricing", href: "#" },
+    ]
+  },
+  {
+    title: "Resources",
+    links: [
+      { label: "Documentation", href: "#" },
+      { label: "Tutorials", href: "#" },
+      { label: "Blog", href: "#" },
+      { label: "Research Papers", href: "#" },
+    ]
+  },
+  {
+    title: "Company",
+    links: [
+      { label: "About Us", href: "/about" },
+      { label: "Careers", href: "#" },
+      { label: "Contact", href: "#" },
+      { label: "Privacy Policy", href: "#" },
+    ]
+  },
+];
