--- qomputing-frontend/utils/constants/homePage.ts
@@ -1,15 +1,0 @@
-// constants/siteContent.ts
-export const heroContent = {
-  tagline: '— Quantum Computing Platform',
-  heading: 'QOMPUTING',
-  subheading: 'Experience the future of computation with our comprehensive quantum development environment. From IDE to simulation, we’ve got your quantum journey covered.',
-  primaryCTA: 'Try Qomputing IDE',
-  secondaryCTA: 'View Documentation',
-};
-
-export const features = [
-  'AI powered Circuit designing',
-  'Interactive UI/UX',
-  'Hardware Integration',
-  'Qomputing® Simulator',
-];
