+++ qomputing-frontend/utils/constants/homePage.tsx
@@ -1,0 +1,90 @@
+export const heroContent = {
+  tagline: '— Quantum Computing Platform',
+  heading: 'QOMPUTING',
+  subheading: 'Experience the future of computation with our comprehensive quantum development environment. From IDE to simulation, we’ve got your quantum journey covered.',
+  primaryCTA: 'Try Qomputing IDE',
+  secondaryCTA: 'View Documentation',
+};
+
+export const heroFeatureList = [
+  'AI powered Circuit designing',
+  'Interactive UI/UX',
+  'Hardware Integration',
+  'Qomputing® Simulator',
+];
+
+export const ctaContent = {
+  heading: 'Ready to Enter the Quantum Era?',
+  subheading: 'Join thousands of developers, researchers, and quantum enthusiasts building the future of computation.',
+  button1: 'START FREE TRIAL',
+  button2: 'JOIN COMMUNITY'
+}
+
+import { Zap, CpuIcon, UsersIcon, BookIcon, WrenchIcon, LineChartIcon } from 'lucide-react';
+
+export const cardTitle = {
+  title: 'Complete Quantum Ecosystem',
+  subheading: 'From development to deployment, our integrated suite of tools empowers you to harness the full potential of quantum computing.'
+}
+export const products = [
+  {
+    icon: <Zap size={32} color="#FCEFB4" />,
+    title: 'Qomputing IDE',
+    points: [
+      'Interactive UI/UX with Drag and Drop feature',
+      'Integrated AI for Circuit Designing',
+      'Integrate with any Quantum Simulator or Hardware',
+      'Our own Qomputing Simulators',
+    ],
+  },
+  {
+    icon: <CpuIcon size={32} color="#FCEFB4" />,
+    title: 'Qomputing Simulator',
+    points: [
+      'Freely available 20+ qubit simulator',
+      'Assembly Language Base',
+      'AI based Qomputing Simulator',
+      'Qomputing Simulators for specific use cases',
+    ],
+  },
+  {
+    icon: <UsersIcon size={32} color="#FCEFB4" />,
+    title: 'Qommunity',
+    points: [
+      'Engage with Similar and Curious minds',
+      'Get updates on Quantum Advancements and research',
+      'Collaborative Research Projects',
+      'Weekly Quantum Workshops & Webinars',
+    ],
+  },
+  {
+    icon: <BookIcon size={32} color="#FCEFB4" />,
+    title: 'Qourses',
+    points: [
+      'Practice on Qomputing IDE',
+      'Hands-on Projects',
+      'Expert Instructors',
+      'Certification',
+    ],
+  },
+  {
+    icon: <WrenchIcon size={32} color="#FCEFB4" />,
+    title: 'Tool Library',
+    points: [
+      'Algorithm Collection',
+      'Research Tools',
+      'Dynamically Typed Codes',
+      'Regular Updates',
+    ],
+  },
+  {
+    icon: <LineChartIcon size={32} color="#FCEFB4" />,
+    title: 'Prediction Modeling',
+    points: [
+      'Monte Carlo Methods',
+      'Risk Quantification',
+      'Probabilistic Models',
+      'Predictive Analytics',
+    ],
+  },
+];
