+++ qomputing-frontend/utils/constants/navbar.ts
@@ -1,0 +1,6 @@
+export const navLinks = [
+  { label: "Products", href: "#" },
+  { label: "Community", href: "#" },
+  { label: "Qourses", href: "#" },
+  { label: "About", href: "/about" },
+];
